% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_to_a.R
\name{aaa_to_a}
\alias{aaa_to_a}
\title{Return one-letter amino acid code from three-letter code}
\usage{
aaa_to_a(aaa)
}
\arguments{
\item{aaa}{AA code (character)}
}
\value{
one-letter AA code (character)
}
\description{
Return one-letter amino acid code from three-letter code
}
\examples{


}
