% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_order.R
\name{new_order}
\alias{new_order}
\title{Reorder vector using factor levels}
\usage{
new_order(vector, order)
}
\arguments{
\item{vector}{vector to be reordered}

\item{order}{vector with new order of elements}
}
\value{
reordered vector of factors
}
\description{
Reorder vector using factor levels
}
\examples{


}
