% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_naccess.R
\name{read_naccess}
\alias{read_naccess}
\title{Read output from Naccess}
\usage{
read_naccess(path)
}
\arguments{
\item{path}{path to .rsa file}
}
\value{
Tibble containing the information from the .rsa file
}
\description{
Read .rsa files and outputs a tibble.
}
\examples{
read_naccess('../naccess/1a2p.rsa')

}
