% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round0.R
\name{round0}
\alias{round0}
\title{Return rounded number as string including trailing zeros}
\usage{
round0(x, decimals)
}
\arguments{
\item{x}{number to round}

\item{y}{number of decimals}
}
\value{
Rounded number
}
\description{
Return rounded number as string including trailing zeros
}
\examples{


}
