% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsquared.R
\name{rsquared}
\alias{rsquared}
\title{Return R squared from linear regression using method "lm"}
\usage{
rsquared(x, y)
}
\arguments{
\item{x}{independent variable vector}

\item{y}{dependent variable vector}
}
\value{
R squared value
}
\description{
Return R squared from linear regression using method "lm"
}
\examples{


}
